requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.5
